requirements.txt
sentence-transformers==3.1.1
dateparser==1.2.0
elasticsearch==8.15.1
flask
flask-cors==5.0.0
eventregistry==9.1
requests==2.31.0
numpy==1.26.4
spacy==3.8.2
thinc==8.3.2
google-generativeai==0.8.3
pandas